feat : 카테고리 숨김 #46
