+++ src/main/java/com/sparta/oneeat/category/entity/Category.java
@@ -3,6 +3,7 @@ package com.sparta.oneeat.category.entity;
 import com.sparta.oneeat.category.dto.CreateCategoryReqDto;
 import com.sparta.oneeat.common.entity.BaseEntity;
 import jakarta.persistence.*;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -24,5 +25,10 @@ public class Category extends BaseEntity {
 
     public Category(CreateCategoryReqDto createCategoryReqDto){
         this.categoryName = createCategoryReqDto.getCategory();
+    }
+
+    public void deleteCategory(long userId) {
+        this.deletedAt = LocalDateTime.now();
+        this.deletedBy = userId;
     }
 }
